ser: implement packing of inner struct
An inner struct is packed with a first block indicating the len of the
whole substruct encoding, followed by this encoding. This is an issue,
because we do not know how much space the substruct encoding will take.

In C, this is done with a first pass over all the substructs, to compute
their relative sizes. This is not possible here, as we want to keep our
packing code derived from every struct definition, and not depending on
reflexive data.

Instead, what is done is the following:
* reserve a block of data in the output, that will contain the tag of
the substruct + the len of its encoding, and save the current len of the
output.
* encode the substruct
* when done, compute the len of the substruct by substracting the len
before its encoding, and save this in the reserved block.

This only issue with this solution is that we have to reserve 4 bytes
all the time, instead of the optimal number of byte for the related len.
To be fair, I'm not sure this is a big loss, compared to the actual size
lookahead computation done in C...