module: add module crate for module require/release
Add a very light shim over C modules, to automate requiring/releasing.
Due to the nature of rust's macros, its not possible to do something
like:

module::require!(ic);

in other crates, so we have to define builders for every crate. This is
done here for the ic crate, and used in tests.